# api/swagger/swagger_ithealth.yaml
openapi: 3.0.1
info:
  title: Swagger IT Health
  description: Documentação da API que será utilizada no aplicativo NINA.
  version: 1.0.0
servers:
  - url: 'http://tstapi.rhhealth.com.br/api/'
paths:
...
  '/usuario/{codigo_usuario}':
    put:
      operationId: putUsuario
      tags:
        - usuario
      description: >-
        Na chamada não pode alterar o login e senha. Caso envie dados no indice
        cliente irá atualizar para o novo cliente.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: codigo_usuario
          schema:
            type: integer
          required: true
          description: ID do usuário ou CPF
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                  example: SERGIO LEANDRO LIMA SALLES
                email:
                  type: string
                  example: teste@teste.com.br
                cpf:
                  type: string
                  example: 07329099708
                data_nascimento:
                  type: string
                  example: '1976-07-05'
                sexo:
                  type: string
                  example: M
                telefone:
                  type: string
                  example: '8138612194'
                celular:
                  type: string
                  example: '81999556553'
                apelido_sistema:
                  type: string
                  example: Serginho
                cliente:
                  type: array
                  items:
                    type: object
                  example: []
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  result:
                    type: object
                    properties:
                      data:
                        type: string
                        example: Registro atualizado com sucesso!
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  result:
                    type: object
                    properties:
                      data:
                        type: string
                        example: Not Found
    get:
      operationId: getUsuario
      tags:
        - usuario
      description: Codigo de usuario para teste 72247
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: codigo_usuario
          schema:
            type: string
          required: true
          description: ID do usuário
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  result:
                    type: object
                    properties:
                      data:
                        type: object
                        properties:
                          codigo_usuario:
                            type: integer
                            example: 134155
                          nome:
                            type: string
                            example: SERGIO LEANDRO LIMA SALLES
                          email:
                            type: string
                            example: teste@teste.com.br
                          celular:
                            type: string
                            example: '81999556553'
                          telefone:
                            type: string
                            example: '8138612194'
                          cpf:
                            type: string
                            example: 07329099708
                          Cliente:
                            type: object
                            properties:
                              nome_fantasia:
                                type: string
                                example: >-
                                  SIEMENS LTDA - SAO PAULO - Anhanguera -
                                  44013159000116
                              razao_social:
                                type: string
                                example: SIEMENS LTDA
                          ContatoEmergencia:
                            type: object
                            properties:
                              nome:
                                type: string
                                example: Will
                              telefone:
                                type: string
                                example: (81) 3861-2194
                              celular:
                                type: string
                                example: (89) 98130-2356
                              grau_parentesco:
                                type: string
                                example: ''
                              email:
                                type: string
                                example: ''
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  result:
                    type: object
                    properties:
                      data:
                        type: string
                        example: Not Found
...
components:
  schemas:
    BadRequest:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: '''CAMPO'' must be a string'
    Unauthorized:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        error:
          type: string
          example: Unauthorized
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT